expose refreshAccessToken as top SDK method to allow client code do access token refresh.
